api: move mailgun to /api/email
Given a domain name like auth-now.sh this will give us the following API
url: `auth-now.sh/api/email/test@domain.com`
